pyproject.toml
[project]
name = "remote_mcp"
version = "0.1.0"
description = "Remote MCP Server Template"
authors = [
    {name = "Nick Armenta", email = "nick@advin.io"},
]

dependencies = [
    "boto3>=1.40.21",
    "mcp[cli]>=1.13.1",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "pytest-xdist>=3.8.0",
    "types-PyYAML>=6.0.0",
    "safety>=3.0.0",
    "bandit[toml]>=1.7.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.11",
    "fastmcp>=2.12.0",
]
